requirements.txt
# Requirements for LLM Benchmark Evaluation Framework

# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# vLLM for acceleration (optional but recommended)
vllm>=0.3.0

# For xVerify evaluation
openai>=1.0.0
tenacity>=8.2.0

# Additional utilities
matplotlib>=3.7.0
seaborn>=0.12.0
